mono repo command
